Refactor: 불필요한 컴포넌트 계층(MembersContent.tsx) 제거
